requirements.txt
Django==1.8.4
django-cms<3.2
django-treebeard>=2.0
pymysql
https://github.com/divio/djangocms-admin-style/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-column/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-flash/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-googlemap/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-inherit/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-link/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-style/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-text-ckeditor/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-file/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-picture/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-teaser/archive/master.zip?1445540826.19
https://github.com/divio/djangocms-video/archive/master.zip?1445540826.19
pytz
django-reversion>=1.8.7
django-classy-tags>=0.3.4.1
html5lib
Pillow>=2.3
django-sekizai>=0.7
six
lxml
gunicorn
requests
python-memcached
